Teach status more options to control which ignored files are shown
The git status command exposes the option to report ignored and
untracked files. When reporting untracked files, it can report untracked
files (--untracked=all), but this results in all ignored files being
reported as well. This teaches Git to optionally show all untracked
files, but not show individual ignored files contained in directories
that match an ignore rule.

Motivation:
Our application (Visual Studio) needs all untracked files listed
individually, but does not need all ignored files listed individually.
Reporting all ignored files can affect the time it takes for status
to run. For a representative repository, here are some measurements
showing a large perf improvement for this scenario:

| Command | Reported ignored entries | Time (s) |
| ------- | ------------------------ | -------- |
| 1       | 0                        | 1.3      |
| 2       | 1024                     | 4.2      |
| 3       | 174904                   | 7.5      |
| 4       | 1046                     | 1.6      |

Commands:
 1) status
 2) status --ignored
 3) status --ignored --untracked-files=all
 4) status --ignored=matching --untracked-files=all

This change adds extra modes to the --ignored option to control which
ignored files are shown. The matching mode is to return ignored files
and directories that match an ignored pattern. In the case where a
directory does not match an ignore pattern, but whose contained paths
are all ignored - the directory will not be shown, but rather all the
contained paths will be shown individually.

Part of the perf improvement comes from a tweak to
read_directory_recursive where it will stop traversing the file system
when it has enough information. This is used when checking whether an
ignored directory is empty or not. read-directory_recursive will stop
when it finds the first file, instead of iterating over all
files. This can result in a perf improvement when an ignored directory
contains a lot of files.

While this change only adds a few modes to the ignored command, this
can be expanded in the future to include the ability to specify --all
and --normal.

Signed-off-by: Jameson Miller <jamill@microsoft.com>