Teach directory logic to efficiently determine if a directory is empty
Introduce the ability for the directory listing logic to determine if a
directory contains any files without recursively iterating through all
files of a directory.

There are scenarios where the directory listing code wants to know if a
directory is empty (e.g. does an ignored folder contain any files). In
order to answer this question, the directory listing logic current recursively iterates
through all files of the directory. This results in unnecessary work as
the algorithm could stop iterating after it encounters any file. In
repositories with a lot of files in an ignored directory stopping after
the first file can result in a noticeable performance increase when
listing ignored files.