Teach directory listing the ability to stop at first file
This commit sets up the ability for the directory listing code to more
efficiently determine if a directory contains any files. Currently, the
directory listing code will iterate through all contents of a directory,
even if it only needs to know if any files exists. This patch adds the
ability to stop recursing after a file is found.

This commit is in preparation for later work which will leverage this
functionality to improve the performance of git status when reporting
ignored files. This can result in a significant performance improvement
in repositories with a lot of files in ignored directories.

This patch series is to start allowing consumers to control which
ignored files are returned independently of which untracked files are
reported. Currently, the ignored files reported by 'git status' is
controlled by which untracked files are reported. It is not possible for
git status to display all untracked files while showing
'normal' ignored files (i.e. displaying all untracked files while
showing ignored directories). If an application wants all untracked
files (for display purposes), but doesn't need all individual ignored
files, it must still pay the price for all ignored files. This can be a
significant amount of extra time when there are many files in ignored
directories.

Our application (Visual Studio) has the requirements to:
 - Get all untracked files individually
 - Get ignored directories / paths