Teach status more options around showing ignored files
Currently, which ignored files are reported by the status command is
controlled by arguments to the `untracked-files` option. This change
teaches the status command more options to control which ignored files
are reported independently of the which untracked files are reported
by allowing the `--ignored` option to take additional arguments.

Our application has a specific set of requirements about how it wants
untracked / ignored files reported by git status. It requires all
untracked files listed individually. It would like ignored files and
directories that explicity match an exclude pattern listed. If an
ignored directory matches an exclude pattern, then the path of the
directory should be returned. If a directory does not match an exclude
pattern, but all of its contents are ignored, then we want the
contained files reported instead of the directory.

The reasons for controlling these behaviors separately is that there
can be a significant performance impact to scanning the contents of
excluded directories. Additionally, knowing whether a directory
explicitly matches an exclude pattern can help the application make
decisions about how to handle the directory. If an ignored directory
itself matches an exclude pattern, then the application will know that
any files underneath the directory must be ignored as well.